// src/features/editor/types.ts
export interface Editor {
	changeSize: (value: { width: number; height: number }) => void;
	changeOpacity: (value: number) => void;
	changeFontSize: (value: number) => void;
	changeTextAlign: (value: string) => void;
	changeFillColor: (value: string) => void;
	changeFontStyle: (value: string) => void;
	changeFontWeight: (value: number) => void;
	changeBackground: (value: string) => void;
	changeFontFamily: (value: string) => void;
	changeImageFilter: (value: string) => void;
	changeStrokeWidth: (value: number) => void;
	changeStrokeColor: (value: string) => void;
	changeFontUnderline: (value: boolean) => void;
	changeFontLinethrough: (value: boolean) => void;
	changeStrokeDashArray: (value: number[]) => void;

	addText: (value: string, options?: ITextboxOptions) => void;
	addImage: (value: string) => void;
	addCircle: () => void;
	addDiamond: () => void;
	addTriangle: () => void;
	addRectangle: () => void;
	addSoftRectangle: () => void;
	addInverseTriangle: () => void;

	getWorkSpace: () => fabric.Object | undefined;
	getActiveOpacity: () => number;
	getActiveFontSize: () => number;
	getActiveTextAlign: () => string;
	getActiveFillColor: () => string;
	getActiveFontStyle: () => string;
	getActiveFontFamily: () => string;
	getActiveFontWeight: () => number;
	getActiveStrokeColor: () => string;
	getActiveStrokeWidth: () => number;
	getActiveFontUnderline: () => boolean;
	getActiveFontLinethrough: () => boolean;
	getActiveStrokeDashArray: () => number[];

	onUndo: () => void;
	onRedo: () => void;
	onCopy: () => void;
	delete: () => void;
	zoomIn: () => void;
	zoomOut: () => void;
	canUndo: () => boolean;
	canRedo: () => boolean;
	onPaste: () => void;
	savePng: (projectId: string) => void;
	saveJpg: (projectId: string) => void;
	saveSvg: (projectId: string) => void;
	saveJson: (projectId: string) => void;
	loadJson: (json: string) => void;
	autoZoom: () => void;
	bringForward: () => void;
	sendBackwards: () => void;
	enableDrawingMode: () => void;
	disableDrawingMode: () => void;

	selectedObjects: fabric.Object[];
	canvas: fabric.Canvas;
}